pyproject.toml
[project]
name = "platform-context-graph"
version = "0.0.3"
description = "An MCP server, CLI, and API for code-to-cloud context graph reasoning by AI assistants."
authors = [{ name = "platformcontext" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "neo4j>=5.15.0",
    "requests>=2.32.0",
    "PyJWT[crypto]>=2.10.1",
    "watchdog>=3.0.0",
    "stdlibs>=2023.11.18",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "inquirerpy>=0.3.4",
    "python-dotenv>=1.0.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-language-pack>=0.6.0",
    "tree-sitter-c-sharp>=0.23.1",
    "pyyaml",
    "pytest",
    "nbformat",
    "nbconvert>=7.16.6",
    "pathspec>=0.12.1",
    "falkordb>=0.1.0",
    "falkordblite>=0.1.0; sys_platform != 'win32' and python_version >= '3.12'",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "httpx>=0.27.0",
    "psycopg[binary]>=3.2.0",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc>=1.34.1",
    "opentelemetry-instrumentation-fastapi>=0.55b1",
]

[project.optional-dependencies]
parsing = [
    "tree-sitter>=0.21.0",
    "tree-sitter-language-pack>=0.6.0",
    "tree-sitter-c-sharp>=0.23.1",
]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
"Homepage" = "https://github.com/platformcontext/platform-context-graph"
"Bug Tracker" = "https://github.com/platformcontext/platform-context-graph/issues"
"Source" = "https://github.com/platformcontext/platform-context-graph"

[project.scripts]
pcg = "platform_context_graph.cli.main:app"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.package-data]
platform_context_graph = ["viz/dist/**/*"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["platform_context_graph*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
norecursedirs = ["tests/fixtures", "tests/e2e/sample_projects", "venv", ".git", "__pycache__"]
markers = [
    "integration: mark a test as an integration test.",
    "e2e: mark a test as an end-to-end test.",
    "slow: mark test as slow."
]

[tool.black]
target-version = ["py310"]
extend-exclude = '''
(
  ^/docs/site/
| ^/src/platform_context_graph\.egg-info/
| ^/tests/fixtures/
)
'''
